+++ unstructured_ingest/processes/connectors/sql/singlestore.py
@@ -102,9 +102,13 @@ class SingleStoreDownloader(SQLDownloader):
         table_name = file_data.additional_metadata.table_name
         id_column = file_data.additional_metadata.id_column
         ids = tuple([item.identifier for item in file_data.batch_items])
-        with self.connection_config.get_connection() as sqlite_connection:
-            cursor = sqlite_connection.cursor()
-            fields = ",".join(self.download_config.fields) if self.download_config.fields else "*"
+        with self.connection_config.get_connection() as connection:
+            cursor = connection.cursor()
+            fields = (
+                ",".join(list(dict.fromkeys([id_column] + self.download_config.fields)))
+                if self.download_config.fields
+                else "*"
+            )
             query = (
                 f"SELECT {fields} FROM {table_name} WHERE {id_column} IN {self.values_delimiter}"
             )